requirements.txt
aiogram==3.13.1
openai>=1.0.0
sentence-transformers==3.3.1
chromadb==0.5.23
aiosqlite==0.20.0
python-dotenv==1.0.1
loguru==0.7.3
aiohttp>=3.9.0
PyMuPDF==1.27.1
